Move griptape.core base classes to matching packages
