+++ griptape/core/adapters/__init__.py
@@ -1,7 +1,9 @@
+from .base_adapter import BaseAdapter
 from .langchain_tool_adapter import LangchainToolAdapter
 from .chatgpt_plugin_adapter import ChatgptPluginAdapter
 
 __all__ = [
+    "BaseAdapter",
     "LangchainToolAdapter",
     "ChatgptPluginAdapter"
 ]
