+++ griptape/core/adapters/chatgpt_plugin_adapter.py
@@ -5,7 +5,8 @@ import yaml
 from attr import define, field
 from fastapi import FastAPI
 from starlette.responses import Response
-from griptape.core import BaseAdapter, BaseTool
+from griptape.core.adapters import BaseAdapter
+from griptape.core.tools import BaseTool
 from griptape.core.utils import J2
