+++ griptape/core/executors/__init__.py
@@ -1,7 +1,9 @@
+from .base_executor import BaseExecutor
 from .local_executor import LocalExecutor
 from .docker_executor import DockerExecutor
 
 __all__ = [
+    "BaseExecutor",
     "LocalExecutor",
     "DockerExecutor"
 ]
