+++ griptape/core/middleware/__init__.py
@@ -1,6 +1,8 @@
+from .base_middleware import BaseMiddleware
 from .storage_middleware import StorageMiddleware
 
 
 __all__ = [
+    "BaseMiddleware",
     "StorageMiddleware"
 ]
