+++ griptape/core/middleware/base_middleware.py
