+++ griptape/utils/tool_loader.py
@@ -1,7 +1,7 @@
 from typing import Optional
 from attr import define, field
-from griptape.core import BaseTool, BaseExecutor
-from griptape.core.executors import LocalExecutor
+from griptape.core.tools import BaseTool
+from griptape.core.executors import BaseExecutor, LocalExecutor
 
 
 @define
