+++ tests/mocks/invalid_mock_tool/tool.py
@@ -1,6 +1,6 @@
 from attr import define, field
 from schema import Schema, Literal
-from griptape.core import BaseTool, action
+from griptape.core.tools import BaseTool, action
 
 
 @define
